Code maintenance

+++ VoxelEngine/src/VoxelEngine/KeyCodes.h
@@ -2,7 +2,7 @@
 
 namespace VoxelEngine
 {
-	using KeyCode = uint16_t;
+	using KeyCode = uint32_t;
 
 	namespace Key
 	{
